backend/requirements.txt
Flask==2.3.3
pymongo==4.6.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
google-generativeai==0.3.2
numpy==1.26.3
python-dotenv==1.0.0